requirements.txt
absl-py==0.9.0
acora==2.2
adversarial-robustness-toolbox==1.10.0
aif360==0.4.0
aiofiles==0.5.0
aiohttp==3.6.2
alabaster==0.7.12
appdirs==1.4.4
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.3.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.8.2
BlackBoxAuditing==0.1.54
bleach==5.0.0
blis==0.4.1
boto3==1.13.6
botocore==1.16.6
cachetools==4.1.0
catalogue==0.2.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
cryptography==36.0.2
cvxpy==1.2.0
cycler==0.11.0
cymem==2.0.3
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
distlib==0.3.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
ecos==2.0.10
# en-core-web-md==2.2.5
entrypoints==0.4
fairlearn==0.4.6
fastjsonschema==2.15.3
fbmessenger==6.0.0
filelock==3.0.12
find-job-titles==0.7.0
Flask==1.1.1
fonttools==4.32.0
future==0.18.2
fuzzywuzzy==0.17.0
gast==0.2.2
gender-guesser==0.4.0
gevent==1.5.0
google-auth==1.14.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GoogleNews==1.3.6
greenlet==0.4.15
grpcio==1.28.1
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.5.5
httplib2==0.17.3
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
idna==2.8
imageio==2.16.2
imagesize==1.3.0
importlib-metadata==1.3.0
importlib-resources==5.7.0
iniconfig==1.1.1
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==3.1.1
jmespath==0.9.5
joblib==0.14.1
jsonpickle==1.4.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.2
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.4.3
lime==0.2.0.1
llvmlite==0.38.0
lxml==4.4.2
Markdown==3.2.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mattermostwrapper==2.2
mccabe==0.6.1
memory-profiler==0.60.0
mistune==0.8.4
mizani==0.7.3
mock==4.0.3
more-itertools==8.0.2
multidict==4.7.5
murmurhash==1.0.2
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
nltk==3.5
notebook==6.4.10
numba==0.55.1
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
osqp==0.6.2.post5
packaging==21.3
palettable==3.3.0
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==9.1.0
pip==20.1.1
pipenv==2020.6.2
plac==1.1.3
plotnine==0.8.0
pluggy==1.0.0
preshed==3.0.2
prometheus-client==0.14.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.9.0
psycopg2-binary==2.8.5
ptyprocess==0.7.0
public==2019.4.13
py==1.11.0
pyahocorasick==1.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydot==1.4.1
Pygments==2.11.2
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.4.4
pymongo==3.8.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.12.1
python-Levenshtein==0.12.0
python-socketio==4.5.1
python-telegram-bot==12.7
pytz==2019.3
PyWavelets==1.3.0
PyYAML==5.3.1
pyzmq==22.3.0
qdldl==0.1.5.post2
qtconsole==5.3.0
QtPy==2.0.1
query-string==2019.4.13
questionary==1.5.2
rasa==1.10.0
rasa-sdk==1.10.0
redis==3.5.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.3.1
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.2
scikit-image==0.19.2
scikit-learn==0.24.2
scipy==1.7.3
scs==3.2.0
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
setuptools==61.2.0
shap==0.40.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==2.5.0
slicer==0.0.7
sniffio==1.1.0
snowballstemmer==2.2.0
soupsieve==1.9.5
spacy==2.2.3
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.16
srsly==1.0.1
statsmodels==0.13.2
tabulate==0.8.7
tempeh==0.1.12
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminado==0.13.3
terminaltables==3.1.0
textblob==0.15.3
thinc==7.3.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
tqdm==4.41.1
traitlets==5.1.1
twilio==6.26.3
typed-ast==1.4.1
typeguard==2.13.3
typing-extensions==4.1.1
tzlocal==2.1
ujson==2.0.3
unicode==2.7
Unidecode==1.1.1
urllib3==1.25.7
#uvloop==0.14.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
wasabi==0.5.0
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.3
websockets==8.1
Werkzeug==1.0.0
wheel==0.34.2
widgetsnbextension==3.6.0
wrapt==1.11.2
yapf==0.29.0
yarl==1.4.2
zipp==0.6.0
